requirements.txt
Flask
Pillow
tensorflow
tensorflow-hub
